[번개] 댓글 삭제 구현완료

+++ your-anmok/src/main/java/com/lion/youranmok/gathering/controller/CommentController.java
@@ -7,6 +7,7 @@ import com.lion.youranmok.gathering.service.GatheringCommentService;
 import com.lion.youranmok.gathering.service.GatheringService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -32,21 +33,24 @@ public class CommentController {
         return "redirect:/gathering/%d".formatted(id);
     }
 
-    @GetMapping("/delete/{id}")
-    public String deleteComment(@PathVariable int id) {
-        GatheringComment gatheringComment = commentService.findById(id);
+    @GetMapping("/delete/{id}/{commentId}")
+    public ResponseEntity deleteComment(@PathVariable int id, @PathVariable int commentId) {
+        GatheringComment gatheringComment = commentService.findById(commentId);
 
         if(gatheringComment == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 댓글이 없습니다.");
         }
-        commentService.delete(gatheringComment);
-
-        return "redirect:/gathering/";
+        if(gatheringComment.getReplyTo() == null) {
+            gatheringComment.setCommentText("삭제된 댓글입니다.");
+            commentService.save(gatheringComment);
+        } else {
+            commentService.delete(gatheringComment);
+        }
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     @PostMapping("/modify/{id}/{commentId}")
     public String modifyComment(@PathVariable int id, @PathVariable int commentId, String content) {
-        System.out.println("content: "+content);
         GatheringComment gatheringComment = commentService.findById(commentId);
         gatheringComment.setCommentText(content);
